+++ Dev/AppUnits/Atdi.AppUnits.Sdrn.DeepServices.RadioSystem/SignalService.cs
@@ -13,7 +13,7 @@ namespace Atdi.AppUnits.Sdrn.DeepServices.RadioSystem
 	public class SignalService : ISignalService
 	{
 
-		public float CalcAntennaGain(in CalcAntennaGainArgs args)
+		public double CalcAntennaGain(in CalcAntennaGainArgs args)
 		{
             return Signal.CalcAntennaGain.Calc(in args);
 		}
